+++ climaf/projects/__init__.py
@@ -45,4 +45,4 @@ from __future__ import print_function, division, absolute_import  # ,unicode_lit
 
 __all__ = ["example", "cami", "cmip3", "cmip5", "ocmip5", "obs4mips", "em", "erai", "erai_land", "gpcc", "gpcp",
            "ceres", "cruts3", "file", "ref_climatos_and_ts", "igcm_out", "emn36", "nemo", "igcm_cmip6", "cmip6", "eobs",
-           "cordex", "cmip6cerfacs", "ref_era5cerfacs"]
+           "cordex", "cmip6cerfacs", "ref_era5cerfacs", "optimize"]