+++ env/environment.py
@@ -60,6 +60,12 @@ locs = list()
 #: Log directory
 logdir = "."
 
+#: Current cache directory
+currentCache = None
+
+#: Should the search for CMI6 files be optimized by building tables
+optimize_cmip6_wildcards = True
+
 #: Define whether we stamp the data files with their CRS.
 # True means mandatory. None means : please try. False means : don't try
 stamping = True