+++ doc/news.rst
@@ -6,6 +6,29 @@ What's new
 
 Changes, newest first:
 
+- V2.0.2:
+
+  - **Finding CMIP6 datasets is optimized when facets/attributes have shell-like wildcards** ( * or ?):
+
+    - This relies on building tables that allows to derive wildcard
+      facet's list of values by using other facet's value. Building
+      the table may take a while, but the table is saved for further
+      sessions (in the cache).
+    - Also, an incremental testing of all possible data paths (built
+      with possible facet values) allows to avoid costly glob.glob
+      operations.
+    - As an example, searching CMIP6 'tos' data for expermient
+      historical and all models, realizations, and grids (and not for
+      peridos) takes 77 seconds on Ciclad;  searching also for
+      all periods takes 20 minutes, instead of 37 minutes.
+    - This is yet limited to CMIP6 data
+    - See module documentation for :py:mod:`~climaf.projects.optimize` 
+
+  - **New dataset's method glob is proposed as a replacement for
+    method explore**. It may be of simpler use, as it can return a list
+    of all facets value combinations that match data. See
+    :py:func:`~climaf.classes.cdataset.glob()`.
+
 - V2.0.1:
 
   - **Scripts can now process multiple ensembles, and ensembles which are not the first argument**: